requirements.txt
fastapi
uvicorn
langchain
google-generativeai
pydantic
python-dotenv
redis
aiofiles
python-multipart